Avoid stdc++ and gcc by linking libconfig++ statically
Since we don't want to ship DLL files for libgcc and libstdc++ and
libconfig seems unable to produce a DLL file where these two are
statically linked, the only way around seems to be to link
libconfig++ statically.

This means we need to ship the full lierolibre source to enable
relinking, even in the binary package.